+++ src/main/java/net/dehydration/mixin/DispenserBlockMixin.java
@@ -1,5 +1,6 @@
 package net.dehydration.mixin;
 
+import net.dehydration.block.RainwaterLeveledCollectorBlock;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
@@ -52,6 +53,13 @@ public abstract class DispenserBlockMixin extends BlockWithEntity {
                     world.setBlockState(newPos, BlockInit.COPPER_WATER_CAULDRON_BLOCK.getDefaultState().with(CopperLeveledCauldronBlock.LEVEL, 3), 3);
                     world.syncWorldEvent(WorldEvents.DISPENSER_DISPENSES, pos, 0);
                     info.cancel();
+                } else if (blockState.isOf(BlockInit.RAINWATER_COLLECTOR_BLOCK)
+                        || (blockState.isOf(BlockInit.RAINWATER_WATER_COLLECTOR_BLOCK) && !((RainwaterLeveledCollectorBlock) blockState.getBlock()).isFull(blockState))) {
+                    itemStack.decrement(1);
+                    dispenserBlockEntity.setStack(i, new ItemStack(Items.BUCKET));
+                    world.setBlockState(newPos, BlockInit.RAINWATER_WATER_COLLECTOR_BLOCK.getDefaultState().with(RainwaterLeveledCollectorBlock.LEVEL, 3), 3);
+                    world.syncWorldEvent(WorldEvents.DISPENSER_DISPENSES, pos, 0);
+                    info.cancel();
                 }
             } else if (itemStack.getItem().equals(ItemInit.PURIFIED_BUCKET)) {
                 if (blockState.isOf(BlockInit.CAMPFIRE_CAULDRON_BLOCK) && blockState.get(CampfireCauldronBlock.LEVEL) < 4) {
@@ -68,6 +76,13 @@ public abstract class DispenserBlockMixin extends BlockWithEntity {
                     world.setBlockState(newPos, BlockInit.COPPER_PURIFIED_WATER_CAULDRON_BLOCK.getDefaultState().with(CopperLeveledCauldronBlock.LEVEL, 3), 3);
                     world.syncWorldEvent(WorldEvents.DISPENSER_DISPENSES, pos, 0);
                     info.cancel();
+                } else if (blockState.isOf(BlockInit.RAINWATER_COLLECTOR_BLOCK)
+                        || (blockState.isOf(BlockInit.RAINWATER_PURIFIED_WATER_COLLECTOR_BLOCK) && !((RainwaterLeveledCollectorBlock) blockState.getBlock()).isFull(blockState))) {
+                    itemStack.decrement(1);
+                    dispenserBlockEntity.setStack(i, new ItemStack(Items.BUCKET));
+                    world.setBlockState(newPos, BlockInit.RAINWATER_PURIFIED_WATER_COLLECTOR_BLOCK.getDefaultState().with(RainwaterLeveledCollectorBlock.LEVEL, 3), 3);
+                    world.syncWorldEvent(WorldEvents.DISPENSER_DISPENSES, pos, 0);
+                    info.cancel();
                 }
             } else if (itemStack.getItem() instanceof PotionItem && itemStack.get(DataComponentTypes.POTION_CONTENTS) != null && itemStack.get(DataComponentTypes.POTION_CONTENTS).potion().isPresent() && itemStack.get(DataComponentTypes.POTION_CONTENTS).potion().get() == ItemInit.PURIFIED_WATER) {
                 if (blockState.isOf(BlockInit.COPPER_CAULDRON_BLOCK)) {
@@ -80,6 +95,18 @@ public abstract class DispenserBlockMixin extends BlockWithEntity {
                     itemStack.decrement(1);
                     dispenserBlockEntity.setStack(i, new ItemStack(Items.GLASS_BOTTLE));
                     world.setBlockState(newPos, (BlockState) blockState.cycle(CopperLeveledCauldronBlock.LEVEL));
+                    world.syncWorldEvent(WorldEvents.DISPENSER_DISPENSES, pos, 0);
+                    info.cancel();
+                }else  if (blockState.isOf(BlockInit.RAINWATER_COLLECTOR_BLOCK)) {
+                    itemStack.decrement(1);
+                    dispenserBlockEntity.setStack(i, new ItemStack(Items.GLASS_BOTTLE));
+                    world.setBlockState(newPos, BlockInit.RAINWATER_PURIFIED_WATER_COLLECTOR_BLOCK.getDefaultState(), 3);
+                    world.syncWorldEvent(WorldEvents.DISPENSER_DISPENSES, pos, 0);
+                    info.cancel();
+                } else if (blockState.isOf(BlockInit.RAINWATER_PURIFIED_WATER_COLLECTOR_BLOCK) && !((RainwaterLeveledCollectorBlock) blockState.getBlock()).isFull(blockState)) {
+                    itemStack.decrement(1);
+                    dispenserBlockEntity.setStack(i, new ItemStack(Items.GLASS_BOTTLE));
+                    world.setBlockState(newPos, (BlockState) blockState.cycle(RainwaterLeveledCollectorBlock.LEVEL));
                     world.syncWorldEvent(WorldEvents.DISPENSER_DISPENSES, pos, 0);
                     info.cancel();
                 }